refactor: 템플릿 컴포넌트 분리 및 컴포넌트 계층화
