+++ src/components/MainPage/RestaurantList.ts
@@ -1,0 +1,12 @@
+import { RestaurantType } from "../../Template";
+import { Restaurant } from "./Restarant";
+
+export const RestaurantList = {
+  template(restaurantList: RestaurantType[]): string {
+    return `<ul class='restaurant-list'>
+        ${restaurantList
+          .map((restaurant) => Restaurant.template(restaurant))
+          .join("")}
+        </ul>`;
+  },
+};
